Split update into predicting and updating.
